[BUGFIX] Prevent quick Asterisk #agi commands from hanging indefinitely by using supported `#execute_agi_command` in Adhearsion.
This change adopts use of execute_agi_command instead of using
Adhearsion::Rayo::Component::Asterisk::AGI::Command as a component, which was a
racey command. Specifically, it executes commands while asyncronously
registering them. Complete Events received from Asterisk for commands that
finish nearly instantly often found the events dropped on the floor, because the
component was not yet registered in the
Adhearsion::Rayo::Client::ComponentRegistry. The outcome in such a case was that
#execute would hang indefinitely, waiting forever for a Complete Event which was
received long ago.